Added the capability to delete old plots.
With the new pool feature in Chia, replotting is necessary. To seamlessly replace old plots with new ones, the script now accepts a third parameter: A directory from which to remove as many files as the number of new plots added.
The assumption here is that the new and the old plots use the same `k` and that the deletion directory only contains plots.